api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Anton Petrov <marblemonolit@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.65.0"
motor = "^2.4.0"
graphene = "^2.1.9"
python-jose = "^3.3.0"
asgi-lifespan = "^1.0.1"
pytz = "^2021.1"
aiokafka = "^0.7.2"
uvicorn = {extras = ["standard"], version = "^0.15.0"}

[tool.poetry.dev-dependencies]
ipdb = "0.13.9"
pytest = "^6.2.4"
pytest-cov = "^2.12.0"
pytest-asyncio = "^0.15.1"
httpx = "^0.18.1"
isort = "^5.9.3"
flake8 = "^3.9.2"
black = "^21.7b0"
Faker = "^8.12.1"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
